+++ SqExpress/IExprSubQuery.cs
@@ -3,7 +3,7 @@ using SqExpress.Syntax.Select;
 
 namespace SqExpress
 {
-    public interface IExprSubQuery : IExprQuery, ISelectingSource
+    public interface IExprSubQuery : IExprReadOnlyQuery, ISelectingSource
     {
         IReadOnlyList<string?> GetOutputColumnNames();
     }